requirements.txt
absl-py==2.0.0
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
audioread==3.0.0
backcall==0.2.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
flatbuffers==20181003210633
fonttools==4.39.3
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
joblib==1.2.0
keras==2.14.0
kiwisolver==1.4.4
lazy-loader==0.2
libclang==16.0.6
librosa==0.10.0.post2
llvmlite==0.39.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mock==5.0.2
msgpack==1.0.5
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
path==16.6.0
path.py==12.5.0
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.5.0
pluggy==1.2.0
pooch==1.6.0
prompt-toolkit==3.0.38
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pytest==7.4.0
pytest-shutil==1.7.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-cpu-aws==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv-clone==0.5.7
wcwidth==0.2.6
werkzeug==3.0.1
wrapt==1.14.1
zipp==3.15.0
